+++ docs/Changelog.md
@@ -3,6 +3,7 @@
 ## New since 0.7.0
 
 - Symbol rotation [#294](https://github.com/mapsforge/vtm/issues/294)
+- Marker clustering [#312](https://github.com/mapsforge/vtm/issues/312)
 - Osmagray theme [#300](https://github.com/mapsforge/vtm/issues/300)
 - OkHttp external cache [#135](https://github.com/mapsforge/vtm/issues/135)
 - Texture atlas improvements [#301](https://github.com/mapsforge/vtm/pull/301) [#304](https://github.com/mapsforge/vtm/pull/304)